Tweets Fetcher
This code will search for all tweets by the specified Twitter account, excluding replies, and extract the Twitter usernames of all followers who are mentioned in those tweets. It will then save the usernames to a CSV file named PlayElarium_followers.csv in the same directory as the script. 
This code will search for all tweets by the specified Twitter account, excluding replies, and extract the Twitter usernames of all followers who are mentioned in those tweets. It will then save the usernames to a CSV file named PlayElarium_followers.csv in the same directory as the script.