Update twitter fetcher
This program uses the snscrape library to scrape tweets from Twitter based on a given query. The query is specified by providing a username, a start date (since_date), and an end date (until_date). The tweets are then saved in a JSON file on the user's desktop, named "tweets.json".

Here's a brief summary of the steps involved in the program:

    Import the required libraries (os, json, datetime, snscrape).
    Specify the output file path (tweets.json) on the user's desktop.
    Specify the query parameters (username, since_date, until_date) for the Twitter search.
    Scrape the tweets based on the query using snscrape.TwitterSearchScraper().
    Filter the tweets by the desired date range and store them in a list.
    Save the list of tweets to the output file using json.dump().